Add files via upload
Long time no see. Rewrote the code for ``ExtractComponents'' and named it FormsToTensors in the fashion of CurvatureFormToTensor etc. The expansion is with the convention I use, this can always be converted to the xAct convention. The idea with the Mark is really nice and can be easily implemented I think. For now, I am using GiveSymbol[Form,Tensor] and clearing/redefining the guy if it already exists.